renaming createFolders by createAllFolders
