+++ docs-generator/create-all-folders.ts
