+++ docs-generator/files-manager/remove-file.ts
@@ -16,6 +16,6 @@ export const removeFile = (filePath: string): void => {
 	if (existsSync(filePath)) {
 		unlinkSync(`${filePath}`);
 	} else {
-		throw `Was not possible to remove this file, ${filePath} does note exists`;
+		console.log(`Was not possible to remove this file, ${filePath} does note exists`);
 	}
 };
